requirements.txt
pandas
google-cloud-storage
moviepy
openpyxl
pydub
pysrt
webvtt-py
datasets
librosa
torch
transformers
jiwer
evaluate
tensorboard
bitsandbytes